% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dict.R
\name{build_sys_dic}
\alias{build_sys_dic}
\title{Build system dictionary}
\usage{
build_sys_dic(dic_dir, out_dir, encoding)
}
\arguments{
\item{dic_dir}{Directory where the source dictionaries are located.
This argument is passed as '-d' option argument.}

\item{out_dir}{Directory where the binary dictionary will be written.
This argument is passed as '-o' option argument.}

\item{encoding}{Encoding of input csv files.
This argument is passed as '-f' option argument.}
}
\value{
A \code{TRUE} is invisibly returned if dictionary is successfully built.
}
\description{
Build a UTF-8 system dictionary from source dictionary files.
}
\details{
This function is a wrapper around dictionary compiler of mecab.

Running this function will create 4 files:
'char.bin', 'matrix.bin', 'sys.dic', and 'unk.dic' in \code{out_dir}.
To use these compiled dictionary, you also need create a \code{dicrc} file to \code{out_dir}.
A \code{dicrc} file is included in source dictionaries, so you can just copy it to \code{out_dir}.
}
