% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dict.R
\name{build_user_dic}
\alias{build_user_dic}
\title{Build user dictionary}
\usage{
build_user_dic(dic_dir, file, csv_file, encoding)
}
\arguments{
\item{dic_dir}{Directory where the source dictionaries are located.
This argument is passed as '-d' option argument.}

\item{file}{Path to write the user dictionary.
This argument is passed as '-u' option argument.}

\item{csv_file}{Path to an input csv file.}

\item{encoding}{Encoding of input csv files.
This argument is passed as '-f' option argument.}
}
\value{
Invisibly returns \code{TRUE} if dictionary is successfully built.
}
\description{
Build a UTF-8 user dictionary from a csv file.
}
\details{
This function is a wrapper around dictionary compiler of mecab.

Note that this function does not support auto assignment of word cost field.
So, you can't leave any word costs as empty in your input csv file.
To assume word costs, use \code{posDebugRcpp()} function.
}
