% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dict_index_sys}
\alias{dict_index_sys}
\title{Build system dictionary}
\arguments{
\item{dic_dir}{Directory where the source dictionaries are located.
This argument is passed as '-d' option argument.}

\item{out_dir}{Directory where the binary dictionary will be written.
This argument is passed as '-o' option argument.}

\item{encoding}{Encoding of input csv files.
This argument is passed as '-f' option argument.}
}
\value{
Logical.
}
\description{
Build system dictionary
}
\keyword{internal}
