% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dict_index_user}
\alias{dict_index_user}
\title{Build user dictionary}
\arguments{
\item{dic_dir}{Directory where the source dictionaries are located.
This argument is passed as '-d' option argument.}

\item{file}{Path to write the user dictionary.
This argument is passed as '-u' option argument.}

\item{csv_file}{Path to an input csv file.}

\item{encoding}{Encoding of input csv files.
This argument is passed as '-f' option argument.}
}
\value{
Logical.
}
\description{
Build user dictionary
}
\keyword{internal}
