% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dictionary_info}
\alias{dictionary_info}
\title{Get dictionary information}
\arguments{
\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}
}
\value{
data.frame.
}
\description{
Get dictionary information
}
