% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-unidic.R
\name{download_unidic}
\alias{download_unidic}
\title{Download and unzip 'UniDic'}
\usage{
download_unidic(version = "latest", dirname = "unidic")
}
\arguments{
\item{version}{String; version of 'UniDic'.}

\item{dirname}{String; directory where unzip the dictionary.}
}
\value{
full path to `dirname` is returned invisibly.
}
\description{
Download 'UniDic' of specified version into `dirname`.
This function is partial port of \href{https://github.com/polm/unidic-py}{polm/unidic-py}.
Note that to unzip dictionary will take up 770MB on disk after downloading.
}
