% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs_c}
\alias{gbs_c}
\title{An alternative of RMeCabC}
\usage{
gbs_c(tbl, pull = "token", names = "POS1")
}
\arguments{
\item{tbl}{A prettified data.frame of tokenized sentences.}

\item{pull}{A column name of `df`.}

\item{names}{A column name of `df`.}
}
\value{
A list of named vectors.
}
\description{
An alternative of RMeCabC
}
