% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs_dfm}
\alias{gbs_dfm}
\title{An alternative of docDF family}
\usage{
gbs_dfm(df, pull = "token", n = 1L, sep = "-", what = "fastestword", ...)
}
\arguments{
\item{df}{A prettified data.frame of tokens.}

\item{pull}{Column to be packed into text or ngrams body. Default value is \code{token}.}

\item{n}{Integer internally passed to ngrams tokenizer function
created of \code{audubon::ngram_tokenizer()}}

\item{sep}{Character scalar internally used as the concatenator of ngrams.}

\item{what}{Character scalar; which tokenizer to use in \code{quanteda::tokens()}.}

\item{...}{Other arguments are passed to \code{quanteda::tokens()}.}
}
\value{
A quanteda 'dfm' object.
}
\description{
Create a sparse document-feature matrix.
This function is a shorthand of
\code{gibasa::gbs_as_tokens()} and then \code{quatenda::dfm()}.
}
\seealso{
Other gbs: 
\code{\link{gbs_as_tokens}()},
\code{\link{gbs_c}()},
\code{\link{gbs_freq}()}
}
\concept{gbs}
