% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs_freq}
\alias{gbs_freq}
\title{An alternative of RMeCabFreq}
\usage{
gbs_freq(df, ..., .name_repair = TRUE)
}
\arguments{
\item{df}{A prettified data.frame of tokenized sentences.}

\item{...}{Other arguments are passed to \code{dplyr::tally()}.}

\item{.name_repair}{Logical:
If true, then rename the column names as RMeCabFreq-compatible style.}
}
\value{
A data.frame.
}
\description{
An alternative of RMeCabFreq
}
\seealso{
Other gbs: 
\code{\link{gbs_as_tokens}()},
\code{\link{gbs_c}()},
\code{\link{gbs_dfm}()}
}
\concept{gbs}
