% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dict_features}
\alias{get_dict_features}
\title{Get features of dictionary}
\usage{
get_dict_features(dict = c("ipa", "unidic17", "unidic26", "unidic29"))
}
\arguments{
\item{dict}{String; one of "ipa", "unidic17", "unidic26" or "unidic29".}
}
\value{
A character vector.
}
\description{
Returns features of dictionary.
Currently supports "unidic17" (2.1.2 src schema), "unidic26" (2.1.2 bin schema),
"unidic29" (schema used in 2.2.0, 2.3.0), and "ipa".
}
