% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_transition_cost}
\alias{get_transition_cost}
\title{Get transition cost between pos attributes}
\usage{
get_transition_cost(rcAttr, lcAttr, sys_dic = "", user_dic = "")
}
\arguments{
\item{rcAttr}{Integer.}

\item{lcAttr}{Integer.}

\item{sys_dic}{String.}

\item{user_dic}{String.}
}
\value{
Numeric.
}
\description{
Get transition cost between pos attributes
}
\keyword{internal}
