% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibasa-package.R
\docType{package}
\name{gibasa-package}
\alias{gibasa}
\alias{gibasa-package}
\title{gibasa: An Alternate 'Rcpp' Interface to 'MeCab'}
\description{
Gibasa is a cleaner 'Rcpp' interface to 'MeCab'.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/paithiov909/gibasa/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
