% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other-reexports.R
\name{lex_density}
\alias{lex_density}
\title{Calculate lexical density}
\usage{
lex_density(vec, contents_words, targets = NULL, negate = c(FALSE, FALSE))
}
\arguments{
\item{vec}{A character vector.}

\item{contents_words}{A character vector containing values
to be counted as contents words.}

\item{targets}{A character vector with which
the denominator of lexical density is filtered before computing values.}

\item{negate}{A logical vector of which length is 2.
If supplied \code{TRUE}, then respectively negates the predicate functions
for counting contents words or targets.}
}
\value{
numeric vector.
}
\description{
The lexical density is the proportion of content words (lexical items)
in documents. This function is a simple helper for calculating
the lexical density of given datasets.
}
