% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pack}
\alias{pack}
\title{Pack prettified data.frame of tokens}
\usage{
pack(df, n = 1L, pull = "token", sep = "-", .collapse = " ")
}
\arguments{
\item{df}{Prettified data.frame of tokens.}

\item{n}{Integer internally passed to ngrams tokenizer function
created of \code{gibasa::ngram_tokenizer()}}

\item{pull}{Column to be packed into text or ngrams body. Default value is `token`.}

\item{sep}{Character scalar internally used as the concatenator of ngrams.}

\item{.collapse}{Character scalar passed to \code{stringi::stri_c()}.}
}
\value{
data.frame
}
\description{
Pack prettified data.frame of tokens into a new data.frame of corpus
compatible with the Text Interchange Formats.
}
\seealso{
\url{https://github.com/ropensci/tif}
}
