% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posDebugRcpp}
\alias{posDebugRcpp}
\title{Tokenizer for debug usage}
\arguments{
\item{text}{String.}

\item{sys_dic}{String.}

\item{user_dic}{String.}
}
\value{
data.frame.
}
\description{
Tokenizes a character vector
and returns all possible results out of the tokenization process.
The returned data.frame contains additional attributes for debug usage.
}
\keyword{internal}
