% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posParallelRcpp}
\alias{posParallelRcpp}
\title{Call tagger inside 'RcppParallel::parallelFor' and return a data.frame.}
\arguments{
\item{text}{Character vector.}

\item{sys_dic}{String scalar.}

\item{user_dic}{String scalar.}

\item{partial}{Logical.}

\item{grain_size}{Integer (larger than 1).}
}
\value{
data.frame.
}
\description{
Call tagger inside 'RcppParallel::parallelFor' and return a data.frame.
}
\keyword{internal}
