% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibasa-package.R
\docType{package}
\name{gibasa-package}
\alias{gibasa}
\alias{gibasa-package}
\title{gibasa: An Alternative 'Rcpp' Interface to 'MeCab'}
\description{
Gibasa is a cleaner 'Rcpp' interface to 'MeCab'.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
