% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibasa-package.R
\docType{package}
\name{gibasa-package}
\alias{gibasa}
\alias{gibasa-package}
\title{gibasa: An Alternate 'Rcpp' Interface to 'MeCab'}
\description{
Gibasa is a plain 'Rcpp' interface to 'MeCab', CJK tokenizer and morphological analysis tool. The main goal of gibasa package is to provide an alternative to 'tidytext' for CJK text data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/gibasa/}
  \item Report bugs at \url{https://github.com/paithiov909/gibasa/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Shogo Ichinose
  \item Taku Kudo
}

Other contributors:
\itemize{
  \item Nippon Telegraph and Telephone Corporation [copyright holder]
}

}
\keyword{internal}
