% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibasa-package.R
\docType{package}
\name{gibasa-package}
\alias{gibasa}
\alias{gibasa-package}
\title{gibasa: An Alternative 'Rcpp' Wrapper of 'MeCab'}
\description{
A plain 'Rcpp' wrapper of 'MeCab' that can segment Chinese, Japanese, and Korean text into tokens. The main goal of this package is to provide an alternative to 'tidytext' using morphological analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paithiov909.github.io/gibasa/}
  \item Report bugs at \url{https://github.com/paithiov909/gibasa/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

Authors:
\itemize{
  \item Shogo Ichinose
  \item Taku Kudo
}

Other contributors:
\itemize{
  \item Jorge Nocedal [contributor]
  \item Nippon Telegraph and Telephone Corporation [copyright holder]
}

}
\keyword{internal}
