% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-rmecab.R
\name{gbs}
\alias{gbs}
\title{Gibasa functions family}
\description{
These functions are compact RMeCab alternatives
based on \href{https://quanteda.io/}{quanteda}.
}
\seealso{
gbs_freq, gbs_c, gbs_as_tokens, gbs_dfm, gbs_collocate
}
\keyword{internal}
