% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_blank}
\alias{is_blank}
\title{Check if scalars are blank}
\usage{
is_blank(x, trim = TRUE, ...)
}
\arguments{
\item{x}{Object to check its emptiness.}

\item{trim}{Logical.}

\item{...}{Additional arguments for \code{base::sapply()}.}
}
\value{
Logicals.
}
\description{
Check if scalars are blank
}
\examples{
is_blank(list(c(a = "", b = NA_character_), NULL))
}
