% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{ngram_tokenizer}
\alias{ngram_tokenizer}
\title{Ngrams tokenizer}
\usage{
ngram_tokenizer(n = 1L)
}
\arguments{
\item{n}{Integer.}
}
\value{
ngram tokenizer function
}
\description{
Makes an ngram tokenizer function.
}
\examples{
bigram <- ngram_tokenizer(2)
bigram(letters, sep = "-")
}
