% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{posParallelRcpp}
\alias{posParallelRcpp}
\title{Call tagger inside 'RcppParallel::parallelFor' and return a data.frame.}
\arguments{
\item{text}{A character vector to be tokenized.}

\item{sys_dic}{Character scalar; path to the system dictionary for 'MeCab'.}

\item{user_dic}{Character scalar; path to the user dictionary for 'MeCab'.}

\item{partial}{Logical; If \code{TRUE}, activates partial parsing mode.}

\item{grain_size}{Integer value larger than 1.}
}
\value{
A data.frame.
}
\description{
This function is an internal function called by \code{tokenize()}.
For common usage, use \code{tokenize()} instead.
}
\keyword{internal}
