% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prettify}
\alias{prettify}
\title{Prettify tokenized output}
\usage{
prettify(df, into = get_dict_features("ipa"))
}
\arguments{
\item{df}{Dataframe that comes out of \code{tokenize()}.}

\item{into}{Character vector that is used as column names of
features.}
}
\value{
data.frame.
}
\description{
Prettify tokenized output
}
