% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize sentence}
\usage{
tokenize(
  sentence,
  sys_dic = "",
  user_dic = "",
  split = FALSE,
  mode = c("parse", "wakati")
)
}
\arguments{
\item{sentence}{Character vector.}

\item{sys_dic}{Character vector; path to the system dictionary for mecab.
Note that the system dictionary is expected to be compiled with UTF-8,
not Shift-JIS or other encodings.}

\item{user_dic}{Character vector; path to the user dictionary for mecab.}

\item{split}{Logical; If true, the function internally split the sentence
into sub-sentences using \code{stringi::stri_split_boudaries(type = "sentence")}.}

\item{mode}{Character scalar to switch output format.}
}
\value{
data.frame or named list.
}
\description{
Tokenize sentence
}
