% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transition_cost}
\alias{transition_cost}
\title{Get transition cost between pos attributes}
\arguments{
\item{rcAttr}{Integer.}

\item{lcAttr}{Integer.}

\item{sys_dic}{String.}

\item{user_dic}{String.}
}
\value{
Numeric.
}
\description{
Get transition cost between pos attributes
}
\keyword{internal}
